Avoiding coupling to domain concepts
