+++ lib/mumukit/login/helpers.rb
@@ -1,3 +1,5 @@
 require_relative './helpers/authentication_helpers'
 require_relative './helpers/authorization_helpers'
 require_relative './helpers/login_controller_helpers'
+require_relative './helpers/organization_helpers'
+require_relative './helpers/login_settings_helpers'
