+++ lib/mumukit/login/helpers/organization_helpers.rb
@@ -1,0 +1,5 @@
+module Mumukit::Login::OrganizationHelpers
+  def login_provider_object
+    @login_provider_object ||= login_provider.try { |it| Mumukit::Login::Provider.parse_login_provider it }
+  end
+end
